Continue tidying sorting.
